Seperate file for WeDevs_Dokan class
